pyproject.toml
[tool.poetry]
name = "bat-stats-api"
version = "0.1.0"
description = ""
authors = ["jghibiki <me@jordan.goetze.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
orjson = "^3.9.10"
pydantic = "^2.5.2"
aiohttp = "^3.9.1"
tdqm = "^0.0.1"
tortoise-orm = {extras = ["asyncpg"], version = "^0.20.0"}
aiohttp-cors = "^0.7.0"
aiocache = {extras = ["redis"], version = "^0.12.2"}



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
